+++ substratevm/src/com.oracle.svm.hosted/src/com/oracle/svm/hosted/NativeImageClassLoaderSupport.java
@@ -77,7 +77,7 @@ import java.util.stream.Stream;
 
 import com.oracle.svm.core.SubstrateUtil;
 import com.oracle.svm.util.ModuleSupport;
-import jdk.graal.nativeimage.LibGraalLoader;
+import org.graalvm.nativeimage.libgraal.LibGraalLoader;
 import org.graalvm.collections.EconomicMap;
 import org.graalvm.collections.EconomicSet;
 import org.graalvm.collections.MapCursor;
